+++ game_patch/misc/alpine_settings.cpp
@@ -292,6 +292,10 @@ bool alpine_player_settings_load(rf::Player* player)
         g_alpine_game_config.gaussian_spread = std::stoi(settings["GaussianSpread"]);
         processed_keys.insert("GaussianSpread");
     }
+    if (settings.count("DisableAllCameraShake")) {
+        g_alpine_game_config.screen_shake_force_off = std::stoi(settings["DisableAllCameraShake"]);
+        processed_keys.insert("DisableAllCameraShake");
+    }
 
     // Load multiplayer settings
     if (settings.count("MultiplayerCharacter")) {
@@ -579,6 +583,7 @@ void alpine_player_settings_save(rf::Player* player)
     file << "DifficultyLevel=" << static_cast<int>(rf::game_get_skill_level()) << "\n";
     file << "UnlimitedSemiAuto=" << g_alpine_game_config.unlimited_semi_auto << "\n";
     file << "GaussianSpread=" << g_alpine_game_config.gaussian_spread << "\n";
+    file << "DisableAllCameraShake=" << g_alpine_game_config.screen_shake_force_off << "\n";
 
     // Multiplayer
     file << "\n[MultiplayerSettings]\n";